rewrite configs
