--- config/production.js
@@ -1,10 +1,0 @@
-export default {
-  host: 'localhost',
-  port: 8080,
-  db: {
-    host: 'mongodb://localhost/anonymClub-prod',
-  },
-  socket: {
-    port: 3000,
-  },
-};
