+++ package.json
@@ -4,10 +4,10 @@
   "description": "anonym.club",
   "scripts": {
     "lint": "eslint .",
-    "dev": "ln -f config/development.js config/index.js; babel-node server/server.dev.js",
-    "prod": "ln -f config/production.js config/index.js; npm run build; cd build; npm run start",
+    "dev": "NODE_ENV=development babel-node server/index.js",
+    "prod": "npm run build; cd build; npm run start",
     "build": "npm run lint; bash build.sh",
-    "test": "ln -f config/testing.js config/index.js; npm run test-server; npm run test-common; npm run test-client",
+    "test": "NODE_ENV=testing npm run test-server; npm run test-common; npm run test-client",
     "test-server": "mocha --compilers js:babel/register 'server/**/*.test.js'",
     "test-common": "mocha --compilers js:babel/register 'common/**/*.test.js'",
     "test-client": "mocha --compilers js:babel/register 'client/**/*.test.js'",
@@ -55,7 +55,9 @@
     "socket.io-client": "1.3.6",
     "style-loader": "0.12.4",
     "webpack": "1.12.2",
-    "webpack-dev-server": "1.11.0"
+    "webpack-dev-middleware": "^1.2.0",
+    "webpack-dev-server": "1.11.0",
+    "webpack-hot-middleware": "^2.4.1"
   },
   "author": "anonym club star team",
   "pre-commit": [
